HSS/requirements.txt
numpy>=1.19.0
opencv-python>=4.2.0
pillow>=7.0.0
pyserial>=3.4
scipy>=1.5.0
pyzbar>=0.1.8 